Flamethrower overhaul!
Flamethrowers no longer mess with atmos or heat whatsoever, instead it just hackily dumps a pile of custom fire objects on the floor and processes them to do burn damage. It also deals serious damage to anything caught directly in the stream, but the leftover flames don't do much actual damage to mobs in them so you can run in and out. Melts alien stuff nicely.